Move the workdir prep into a BuildStage class
