+++ dockci/models/build_meta/__init__.py
@@ -1,0 +1,3 @@
+"""
+Metadata that helps to describe builds
+"""
