+++ dockci/models/build_meta/stages_main.py
@@ -1,0 +1,3 @@
+"""
+Main build stages that constitute a build
+"""
