+++ pylint.conf
@@ -8,6 +8,7 @@ disable=bad-continuation,
         locally-enabled,
         relative-import,
         star-args,
+        too-few-public-methods,
         wildcard-import,
 
 [REPORTS]
